+++ src/main/java/rbasamoyai/createbigcannons/datagen/assets/CBCBuilderTransformers.java
@@ -508,6 +508,7 @@ public class CBCBuilderTransformers {
 						.when(LootItemBlockStatePropertyCondition.hasBlockStateProperties(c)
 							.setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(BigCannonMunitionBlock.DAMP, false)))));
 			})
+            .tag(AllBlockTags.SAFE_NBT.tag)
 			.item(BigCartridgeBlockItem::new)
 			.tag(CBCTags.CBCItemTags.BIG_CANNON_CARTRIDGES)
 			.model((c, p) -> {