+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Fixed:
 - Fixed Create contraption attachment of Cannon Drill and Cannon Builder
 - Fixed cannon shots not hitting entities if they are destroyed from hitting a block
 - Fixed flak, shrapnel, and grapeshot burst overpenetration
+- Fixed schematic requirements for big cartridges and big cannon projectiles
 - [Aeronautics] Fixed cannon blocks and cannon contraptions not attaching to ship on assembly
 - [Aeronautics] Fixed Cannon Loader, Cannon Drill, and Cannon Builder not attaching to ship on assembly
 - [Sable] Fixed cannon blocks splitting from cannon mount on physics object